+++ pkg/contracts/solana/gateway_message.go
@@ -17,6 +17,17 @@ const (
 	InstructionExecute           byte = 5
 	InstructionExecuteSPL        byte = 6
 	InstructionIncrementNonce    byte = 7
+	InstructionExecuteRevert     byte = 8
+	InstructionExecuteSPLRevert  byte = 9
+)
+
+// ExecuteType represents the type of execute operation
+// it is used for withdraw and call and revert
+type ExecuteType uint8
+
+const (
+	ExecuteTypeCall ExecuteType = iota
+	ExecuteTypeRevert
 )
 
 // InstructionIdentifier is used at beginning of message hash to make it project specific.
@@ -231,10 +242,13 @@ type MsgExecute struct {
 	signature [65]byte
 
 	// Sender is the sender address for the execute
-	sender [20]byte
+	sender string
 
 	// Data for execute
 	data []byte
+
+	// executeType indicates if it's a call or revert operation
+	executeType ExecuteType
 
 	// Remaining accounts for arbirtrary program
 	remainingAccounts []*solana.AccountMeta
@@ -244,8 +258,9 @@ type MsgExecute struct {
 func NewMsgExecute(
 	chainID, nonce, amount uint64,
 	to solana.PublicKey,
-	sender [20]byte,
+	sender string,
 	data []byte,
+	executeType ExecuteType,
 	remainingAccounts []*solana.AccountMeta,
 ) *MsgExecute {
 	return &MsgExecute{
@@ -255,6 +270,7 @@ func NewMsgExecute(
 		to:                to,
 		sender:            sender,
 		data:              data,
+		executeType:       executeType,
 		remainingAccounts: remainingAccounts,
 	}
 }
@@ -280,7 +296,7 @@ func (msg *MsgExecute) To() solana.PublicKey {
 }
 
 // Sender returns the sender address of the message
-func (msg *MsgExecute) Sender() [20]byte {
+func (msg *MsgExecute) Sender() string {
 	return msg.sender
 }
 
@@ -294,13 +310,23 @@ func (msg *MsgExecute) RemainingAccounts() []*solana.AccountMeta {
 	return msg.remainingAccounts
 }
 
+// ExecuteType returns the type of execute operation
+func (msg *MsgExecute) ExecuteType() ExecuteType {
+	return msg.executeType
+}
+
 // Hash packs the execute message and computes the hash
 func (msg *MsgExecute) Hash() [32]byte {
 	var message []byte
 	buff := make([]byte, 8)
 
 	message = append(message, InstructionIdentifier...)
-	message = append(message, InstructionExecute)
+
+	if msg.executeType == ExecuteTypeCall {
+		message = append(message, InstructionExecute)
+	} else {
+		message = append(message, InstructionExecuteRevert)
+	}
 
 	binary.BigEndian.PutUint64(buff, msg.chainID)
 	message = append(message, buff...)
@@ -507,10 +533,13 @@ type MsgExecuteSPL struct {
 	signature [65]byte
 
 	// Sender is the sender address for the execute spl
-	sender [20]byte
+	sender string
 
 	// Data for execute
 	data []byte
+
+	// executeType indicates if it's a call or revert operation
+	executeType ExecuteType
 
 	// Remaining accounts for arbirtrary program
 	remainingAccounts []*solana.AccountMeta
@@ -521,8 +550,9 @@ func NewMsgExecuteSPL(
 	chainID, nonce, amount uint64,
 	decimals uint8,
 	mintAccount, to, toAta solana.PublicKey,
-	sender [20]byte,
+	sender string,
 	data []byte,
+	executeType ExecuteType,
 	remainingAccounts []*solana.AccountMeta,
 ) *MsgExecuteSPL {
 	return &MsgExecuteSPL{
@@ -535,6 +565,7 @@ func NewMsgExecuteSPL(
 		decimals:          decimals,
 		sender:            sender,
 		data:              data,
+		executeType:       executeType,
 		remainingAccounts: remainingAccounts,
 	}
 }
@@ -572,7 +603,7 @@ func (msg *MsgExecuteSPL) Decimals() uint8 {
 }
 
 // Sender returns the sender address of the message
-func (msg *MsgExecuteSPL) Sender() [20]byte {
+func (msg *MsgExecuteSPL) Sender() string {
 	return msg.sender
 }
 
@@ -586,13 +617,23 @@ func (msg *MsgExecuteSPL) RemainingAccounts() []*solana.AccountMeta {
 	return msg.remainingAccounts
 }
 
+// ExecuteType returns the type of execute operation
+func (msg *MsgExecuteSPL) ExecuteType() ExecuteType {
+	return msg.executeType
+}
+
 // Hash packs the execute spl message and computes the hash
 func (msg *MsgExecuteSPL) Hash() [32]byte {
 	var message []byte
 	buff := make([]byte, 8)
 
 	message = append(message, InstructionIdentifier...)
-	message = append(message, InstructionExecuteSPL)
+
+	if msg.executeType == ExecuteTypeCall {
+		message = append(message, InstructionExecuteSPL)
+	} else {
+		message = append(message, InstructionExecuteSPLRevert)
+	}
 
 	binary.BigEndian.PutUint64(buff, msg.chainID)
 	message = append(message, buff...)