+++ x/crosschain/keeper/evm_hooks.go
@@ -261,7 +261,8 @@ func (k Keeper) ProcessZetaSentEvent(
 		return observertypes.ErrChainParamsNotFound
 	}
 
-	if receiverChain.IsExternalChain() && chainParams.ZetaTokenContractAddress == "" {
+	if receiverChain.IsExternalChain() &&
+		(chainParams.ZetaTokenContractAddress == "" || chainParams.ZetaTokenContractAddress == constant.EVMZeroAddress) {
 		return types.ErrUnableToSendCoinType
 	}
 
@@ -317,12 +318,12 @@ func (k Keeper) ValidateZRC20WithdrawEvent(
 	coinType coin.CoinType,
 ) error {
 	// The event was parsed; that means the user has deposited tokens to the contract.
-	return k.validateZRC20Withdrawal(ctx, chainID, coinType, event.Value, event.To)
+	return k.validateOutbound(ctx, chainID, coinType, event.Value, event.To)
 }
 
-// validateZRC20Withdrawal validates the data of a ZRC20 Withdrawal event (version 1 or 2)
+// validateOutbound validates the data of a ZRC20 Withdrawals and Call event (version 1 or 2)
 // it checks if the withdrawal amount is valid and the destination address is supported depending on the chain
-func (k Keeper) validateZRC20Withdrawal(
+func (k Keeper) validateOutbound(
 	ctx sdk.Context,
 	chainID int64,
 	coinType coin.CoinType,
@@ -348,7 +349,8 @@ func (k Keeper) validateZRC20Withdrawal(
 		}
 	} else if chains.IsSolanaChain(chainID, additionalChains) {
 		// The rent exempt check is not needed for ZRC20 (SPL) tokens because withdrawing SPL token
-		// already needs a non-trivial amount of SOL for potential ATA creation so we can skip the check.
+		// already needs a non-trivial amount of SOL for potential ATA creation so we can skip the check,
+		// and also not needed for simple no asset call.
 		if coinType == coin.CoinType_Gas && value.Cmp(big.NewInt(constant.SolanaWalletRentExempt)) < 0 {
 			return errorsmod.Wrapf(
 				types.ErrInvalidWithdrawalAmount,